SA-#138 Fixed bug: joined player did not have opportunity to vote

+++ lib/services/score_service.dart
@@ -56,7 +56,7 @@ class ScoreService {
   Future<void> deleteOldScore(DocumentReference participantRef, int questionNr) async {
     QuerySnapshot score =
         await participantRef.collection(SCORES_SUB_COLLECTION).where(QUESTION, isEqualTo: questionNr).get();
-    return score.docs.single.reference.delete();
+    if (score.docs.isNotEmpty) return score.docs.single.reference.delete();
   }
 
   Future<void> updateAgreedScore(DocumentReference gameRef, int agreedScore, int questionNr) async {
